+++ programs/su/CMakeLists.txt
@@ -1,3 +1,3 @@
-set (CLICKHOUSE_SU_SOURCES clickhouse-su.cpp)
+set (CLICKHOUSE_SU_SOURCES clickhouse-su.cpp su.cpp)
 set (CLICKHOUSE_SU_LINK PRIVATE dbms)
 clickhouse_program_add(su)
